requirements.txt
langchain
langchain_groq
groq 
streamlit
python-dotenv
